+++ dart/geocore/lib/src/base/multi.dart
@@ -50,8 +50,8 @@ class GeometryCollection<T extends Geometry> extends Geometry
   List<Object?> get props => [geometries];
 
   @override
-  GeometryCollection<T> project(TransformPoint transform) =>
-      GeometryCollection(geometries.project(transform, lazy: false));
+  GeometryCollection<T> transform(TransformPoint transform) =>
+      GeometryCollection(geometries.transform(transform, lazy: false));
 }
 
 /// A multi point geometry.
@@ -101,8 +101,8 @@ class MultiPoint<T extends Point> extends Geometry with EquatableMixin {
   List<Object?> get props => [points];
 
   @override
-  MultiPoint<T> project(TransformPoint transform) =>
-      MultiPoint(points.project(transform, lazy: false));
+  MultiPoint<T> transform(TransformPoint transform) =>
+      MultiPoint(points.transform(transform, lazy: false));
 }
 
 /// A multi line string geometry.
@@ -164,8 +164,8 @@ class MultiLineString<T extends Point> extends Geometry with EquatableMixin {
   List<Object?> get props => [lineStrings];
 
   @override
-  MultiLineString<T> project(TransformPoint transform) =>
-      MultiLineString(lineStrings.project(transform, lazy: false));
+  MultiLineString<T> transform(TransformPoint transform) =>
+      MultiLineString(lineStrings.transform(transform, lazy: false));
 }
 
 /// A multi polygon geometry.
@@ -226,6 +226,6 @@ class MultiPolygon<T extends Point> extends Geometry with EquatableMixin {
   List<Object?> get props => [polygons];
 
   @override
-  MultiPolygon<T> project(TransformPoint transform) =>
-      MultiPolygon(polygons.project(transform, lazy: false));
+  MultiPolygon<T> transform(TransformPoint transform) =>
+      MultiPolygon(polygons.transform(transform, lazy: false));
 }