+++ dart/geocore/lib/src/base/point.dart
@@ -175,12 +175,21 @@ abstract class Point<C extends num> extends Geometry
   /// value, then it's ignored.
   Point copyWith({num? x, num? y, num? z, num? m});
 
-  /// Returns a new point projected from this point using [transform].
+  /// Returns a new point transformed from this point using [transform].
   ///
-  /// The projected point object must be of the type with same coordinate value
-  /// members as this object has.
+  /// The transformed point object must be of the type with same coordinate
+  /// value members as this object has.
   @override
-  Point project(TransformPoint transform);
+  Point transform(TransformPoint transform);
+
+  /// Returns a new point projected from this point using [project] function.
+  ///
+  /// When [to] is provided, then target points of [R] are created using
+  /// that as a point factory. Otherwise [project] uses it's own factory.
+  R project<R extends Point>(
+    ProjectPoint<Point, R> project, {
+    PointFactory<R>? to,
+  });
 }
 
 /// A read-only cartesian (or projected) point with coordinate value getters.
@@ -201,7 +210,14 @@ abstract class CartesianPoint<C extends num> extends Point<C> {
   CartesianPoint newFrom(Iterable<num> coords, {int? offset, int? length});
 
   @override
-  CartesianPoint project(TransformPoint transform);
+  CartesianPoint transform(TransformPoint transform);
+
+  @override
+  R project<R extends Point>(
+    ProjectPoint<CartesianPoint, R> project, {
+    PointFactory<R>? to,
+  }) =>
+      project(this, to: to);
 }
 
 /// A private implementation for an empty point with coordinate zero values.
@@ -250,7 +266,14 @@ class _PointEmpty<C extends num> extends Point<C> with EquatableMixin {
   Point copyWith({num? x, num? y, num? z, num? m}) => this;
 
   @override
-  Point project(TransformPoint transform) => this;
+  Point transform(TransformPoint transform) => this;
+
+  @override
+  R project<R extends Point>(
+    ProjectPoint<Point, R> project, {
+    PointFactory<R>? to,
+  }) =>
+      throw const FormatException('Cannot project empty point.');
 
   @override
   List<Object?> get props => [is3D, hasM];