+++ dart/geocore/lib/src/geo.dart
@@ -1,3 +1,4 @@
 export 'geo/geobounds.dart';
 export 'geo/geopoint.dart';
 export 'geo/geopoint_wrapper.dart';
+export 'geo/webmercator.dart';