+++ dart/geodata/lib/src/api/features/feature_items.dart
@@ -36,6 +36,6 @@ class FeatureItems<T extends Feature> extends FeatureCollection<T>
   Bounds get bounds => collection.bounds;
 
   @override
-  FeatureItems<T> project(TransformPoint transform) =>
-      FeatureItems(collection: collection.project(transform), meta: meta);
+  FeatureItems<T> transform(TransformPoint transform) =>
+      FeatureItems(collection: collection.transform(transform), meta: meta);
 }