+++ dart/geocore/CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Test point equality in 2D or 3D by tolerance in Point class #47](https://github.com/navibyte/geospatial/issues/47)
 - [Check if point series is closed by tolerance #48](https://github.com/navibyte/geospatial/issues/48)
 - [Add CartesianPoint to the Point-class hierarchy #54](https://github.com/navibyte/geospatial/issues/54)
+- [Coordinate conversion between WGS84 (lon-lat) and Web Mercator projection (meters)](https://github.com/navibyte/geospatial/issues/55)
 
 🛠 Refactoring:
 - [Refactor coordinate value members of immutable Point classes #44](https://github.com/navibyte/geospatial/issues/44)